Warning fixes. Support for >4gb in GFile::Read/Write.

+++ include/common/GCss.h
@@ -803,7 +803,7 @@ public:
 		{
 			const char *Element = Context->GetElement(Obj);
 			
-			for (ssize_t n = PartIdx; n<Sel->Parts.Length(); n++)
+			for (size_t n = PartIdx; n<Sel->Parts.Length(); n++)
 			{
 				GCss::Selector::Part &p = Sel->Parts[n];
 				switch (p.Type)
@@ -941,7 +941,7 @@ public:
 					StartIdx = Sel->Combs[CombIdx];
 					LgiAssert(StartIdx > 0);
 
-					if (StartIdx >= Sel->Parts.Length())
+					if (StartIdx >= (ssize_t)Sel->Parts.Length())
 						break;
 					
 					GCss::Selector::Part &p = Sel->Parts[StartIdx];
